Improve tests: symlink tests, Kotlin tests, and maybe test-util module (#2605)
* Ignore Kotlin tests and show messages when Java is not installed or configured. (#2603)

* Skip symlink tests and show warning message when tests are run without admin privilege. (#2602)

* Moved parcel-bundler/test/utils.js and integration-tests/test/utils.js into test-utils/src/utils.js. (#2604)

* Fixed eslint errors.

* Use the test to assert this.child.killed rather than checking time difference. (#2609) (#2612)
