+++ packages/core/integration-tests/test/contentHashing.js
@@ -1,7 +1,7 @@
 const assert = require('assert');
 const path = require('path');
 const fs = require('@parcel/fs');
-const {bundle, rimraf, ncp} = require('./utils');
+const {bundle, rimraf, ncp} = require('@parcel/test-utils');
 
 describe('content hashing', function() {
   beforeEach(async function() {
