+++ packages/core/integration-tests/test/glsl.js
@@ -1,7 +1,12 @@
 const assert = require('assert');
 const path = require('path');
 const fs = require('@parcel/fs');
-const {bundle, run, assertBundleTree, normaliseNewlines} = require('./utils');
+const {
+  bundle,
+  run,
+  assertBundleTree,
+  normaliseNewlines
+} = require('@parcel/test-utils');
 
 describe('glsl', function() {
   it('should support requiring GLSL files via glslify', async function() {
