+++ packages/core/integration-tests/test/server.js
@@ -2,7 +2,7 @@ const assert = require('assert');
 const path = require('path');
 const fs = require('@parcel/fs');
 const logger = require('@parcel/logger');
-const {bundler} = require('./utils');
+const {bundler} = require('@parcel/test-utils');
 const http = require('http');
 const https = require('https');
 const sinon = require('sinon');
