+++ packages/core/parcel-bundler/test/resolver.js
@@ -1,35 +1,45 @@
 const Resolver = require('../src/Resolver');
 const path = require('path');
 const assert = require('assert');
-const {rimraf, ncp} = require('./utils');
+const {rimraf, ncp} = require('@parcel/test-utils');
 const {mkdirp} = require('@parcel/fs');
+const {symlinkPrivilegeWarning} = require('@parcel/test-utils');
 const {symlinkSync} = require('fs');
 
 const rootDir = path.join(__dirname, 'input/resolver');
 
 describe('resolver', function() {
   let resolver;
+  let hasPrivilege = true;
+
   before(async function() {
     await rimraf(path.join(__dirname, '/input'));
     await mkdirp(rootDir);
     await ncp(path.join(__dirname, 'integration/resolver'), rootDir);
 
     // Create the symlinks here to prevent cross platform and git issues
-    symlinkSync(
-      path.join(rootDir, 'packages/source'),
-      path.join(rootDir, 'node_modules/source'),
-      'dir'
-    );
-    symlinkSync(
-      path.join(rootDir, 'packages/source-alias'),
-      path.join(rootDir, 'node_modules/source-alias'),
-      'dir'
-    );
-    symlinkSync(
-      path.join(rootDir, 'packages/source-alias-glob'),
-      path.join(rootDir, 'node_modules/source-alias-glob'),
-      'dir'
-    );
+    try {
+      symlinkSync(
+        path.join(rootDir, 'packages/source'),
+        path.join(rootDir, 'node_modules/source'),
+        'dir'
+      );
+      symlinkSync(
+        path.join(rootDir, 'packages/source-alias'),
+        path.join(rootDir, 'node_modules/source-alias'),
+        'dir'
+      );
+      symlinkSync(
+        path.join(rootDir, 'packages/source-alias-glob'),
+        path.join(rootDir, 'node_modules/source-alias-glob'),
+        'dir'
+      );
+    } catch (e) {
+      if (e.code == 'EPERM') {
+        symlinkPrivilegeWarning();
+        hasPrivilege = false;
+      }
+    }
 
     resolver = new Resolver({
       rootDir,
@@ -579,6 +589,8 @@ describe('resolver', function() {
 
   describe('source field', function() {
     it('should use the source field when symlinked', async function() {
+      if (!hasPrivilege) this.skip();
+
       let resolved = await resolver.resolve(
         'source',
         path.join(rootDir, 'foo.js')
@@ -591,6 +603,8 @@ describe('resolver', function() {
     });
 
     it('should not use the source field when not symlinked', async function() {
+      if (!hasPrivilege) this.skip();
+
       let resolved = await resolver.resolve(
         'source-not-symlinked',
         path.join(rootDir, 'foo.js')
@@ -603,6 +617,8 @@ describe('resolver', function() {
     });
 
     it('should use the source field as an alias when symlinked', async function() {
+      if (!hasPrivilege) this.skip();
+
       let resolved = await resolver.resolve(
         'source-alias/dist',
         path.join(rootDir, 'foo.js')
@@ -615,6 +631,8 @@ describe('resolver', function() {
     });
 
     it('should use the source field as a glob alias when symlinked', async function() {
+      if (!hasPrivilege) this.skip();
+
       let resolved = await resolver.resolve(
         'source-alias-glob',
         path.join(rootDir, 'foo.js')
